backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["shadowik <zahar.mark@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
alembic = "^1.14.1"
pydantic-core = "^2.29.0"
pydantic-settings = "^2.7.1"
pyjwt = "^2.10.1"
bcrypt = "^4.2.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
black = "^25.1.0"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
